+++ lib/flutter_constraintlayout.dart
@@ -2,4 +2,4 @@ library flutter_constraintlayout;
 
 export 'src/core.dart';
 export 'src/extensions.dart';
-export 'src/utils.dart';
+export 'src/auxiliary.dart';